+++ packages/graphql/tests/e2e/subscriptions/delete-relationship-via-delete.e2e.test.ts
@@ -23,11 +23,11 @@ import { Neo4jGraphQL } from "../../../src/classes";
 import { UniqueType } from "../../utils/graphql-types";
 import type { TestGraphQLServer } from "../setup/apollo-server";
 import { ApolloTestServer } from "../setup/apollo-server";
-import { TestSubscriptionsEngine } from "../../utils/TestSubscriptionsEngine";
 import { WebSocketTestClient } from "../setup/ws-client";
 import Neo4j from "../setup/neo4j";
 import { cleanNodes } from "../../utils/clean-nodes";
 import { delay } from "../../../src/utils/utils";
+import { Neo4jGraphQLSubscriptionsDefaultEngine } from "../../../src/classes/subscription/Neo4jGraphQLSubscriptionsDefaultEngine";
 
 describe("Delete Subscriptions when only nodes are targeted - with interfaces, unions and label manipulation", () => {
     let neo4j: Neo4j;
@@ -98,7 +98,7 @@ describe("Delete Subscriptions when only nodes are targeted - with interfaces, u
             typeDefs,
             driver,
             features: {
-                subscriptions: new TestSubscriptionsEngine(),
+                subscriptions: new Neo4jGraphQLSubscriptionsDefaultEngine(),
             },
         });
         // eslint-disable-next-line @typescript-eslint/require-await